(SIMP-1410) Updated the RPM spec template
* Fixed the template so that it properly brings in the 'requires' and
  'release' files from the RPM metadata directory
* Ensure that the appropriate files are copied into the 'dist'
  directory so that they make their way into the SOURCES directory
* This needs to be scrapped and redone more cleanly in the future. Got
  too clever with the LUA-fu

SIMP-1410 #comment Fixed the RPM spec template

+++ lib/simp/rake/pkg.rb
@@ -267,9 +267,15 @@ module Simp::Rake
             if require_rebuild?(@tar_dest,srpms) || require_rebuild?("#{@base_dir}/metadata.json")
 
               @puppet_module_info_files.each do |file|
-                tgt_file = File.join(@pkg_dir, File.basename(file))
-                FileUtils.rm_rf(tgt_file) if File.exist?(tgt_file)
-                FileUtils.cp_r(file, @pkg_dir) if File.exist?(file)
+                next unless File.exist?(file)
+
+                Find.find(file) do |path|
+                  next if File.directory?(path)
+
+                  tgt_file = File.join(@pkg_dir, File.basename(path))
+                  FileUtils.rm_rf(tgt_file) if File.exist?(tgt_file)
+                  FileUtils.cp(path, @pkg_dir) if File.exist?(path)
+                end
               end
 
               cmd = %Q(#{mock_cmd} --root #{args.chroot} #{mocksnap} --buildsrpm --spec #{@spec_file} --sources #{@pkg_dir})
